MGS-3214 [#ccc] Integrate 6.2.4 official release
Support GEM feature for Android DRM and X11 DRI3,
Fixed the OpenVX 1.1 CTS failures for GC7000XSVX,
Add performance optimization for GPU benchmarking.

Release package from:
Verisilicon_SW_GAL_Unified_Src_drv_6_2_4_NXP_20171031

Rebase new package to fix copyright:
Verisilicon_SW_GAL_Unified_Src_drv_6_2_4_NUpdate_20171102

Date: Nov 02, 2017

+++ driver/driver/khronos/libGL4/src/glcore/display_list/gc_gl_dl_vfcull.c
@@ -530,7 +530,7 @@ Restart:
             for (index = 0; index < newBegin->indexCount; index++)
             {
                 gc->input.defaultIndexBuffer[primBegin->indexCount + index] =
-                                        newBegin->indexBuffer[index] + primBegin->vertexCount;
+                                        newBegin->indexBuffer[index] + (GLushort)primBegin->vertexCount;
             }
         }
         else
@@ -542,7 +542,7 @@ Restart:
                 for (index = 0; index < newBegin->indexCount; index++)
                 {
                     gc->input.defaultIndexBuffer[primBegin->indexCount + index] =
-                                            newBegin->indexBuffer[index] + primBegin->vertexCount;
+                                            newBegin->indexBuffer[index] + (GLushort)primBegin->vertexCount;
                 }
             }
             if (newBegin->indexCount == 0 && newBegin->primType == GL_TRIANGLES)
